rosplan_knowledge_base: fix bug misidentifying duplicates by comparitor
The comparitor required only a single argument to a knowledge item to be
the same to consider the overall item the same. Fix the condition such
that any difference correctly identifies the full item as different.

Note that this bug was not introduced by the case-insensitive change but
existed before.