Minor Change in Predicate Extraction on Parameters
This is a minor modification to the CFFPlanParser to include more general cases covered by PDDL 2.1. The planner was able to give a valid plan but the parser would not work correctly. The earliest version identifies the predicate by locating its type. So it always should have a type after every predicate, in other to work. 
For example it would work in:
:parameters (?q - robot ?wayp - waypoint ?wayp2 - waypoint )

But not in:
:parameters (?q - robot ?wayp ?wayp2 - waypoint )   or   :parameters (?q ?wayp ?wayp2) 

The proposed version uses the same original structure but identify the predicate by locating the "?"