Update CFFPlanParser.cpp
This is a minor modification to the CFFPlanParser to include more general cases covered by PDDL 2.1. The planner was able to give a valid plan but the parser would not work correctly. The earliest version identifies the parameter by locating its type. So it always should have a type after every parameter, in other to work. 
For example it would work in:
:parameters (?q - robot ?wayp - waypoint ?wayp2 - waypoint )

But not in:
:parameters (?q - robot ?wayp ?wayp2 - waypoint )   or   :parameters (?q ?wayp ?wayp2) 

The proposed version uses the same original structure but identify the parameter by locating the "?"