Add problem generation for CHIMP
The CHIMPProblemGenerator queries the knowledge base for all relevant information
which it adds to CHIMPProblem. The CHIMPProblem collects all necessary information
and provides a method for generating the problem file in CHIMP's format.

The ProblemInterface is adapted such that it uses the CHIMPProblemGenerator when the
problem file as the ending ".pdl" or the planning language is set to "chimp".